requirements.txt
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
docutils==0.21.2
filetype==1.2.0
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
Pygments==2.19.1
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==308
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
